Allow custom theming. :art:
