+++ src/components/Media/Media.styled.tsx
@@ -1,0 +1,11 @@
+import { styled } from "@stitches/react";
+
+const MediaWrapper = styled("nav", {
+  display: "flex",
+  flexDirection: "row",
+  flexGrow: "1",
+  padding: "1.618rem 0.618rem 1.618rem 0 ",
+  overflowX: "scroll",
+});
+
+export { MediaWrapper };
